JIRA: RW-2559 Eliminate compiler warnings in alignment reader code. AUTHOR: foleyjp MERGE_REQUEST:  586556f495589125511e5338bbdd22e2443ca0a5
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@105068 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/aln_util.hpp
@@ -67,7 +67,7 @@ namespace AlnUtil
     void CheckId(
         const string& seqId,
         const vector<SLineInfo>& orderedIds,
-        int idCount,
+        size_t idCount,
         int lineNum,
         bool firstBlock);
